Extracts spy field initialization code
